Prepare test framework
