+++ Tests/FuturaTests/Future/PromiseAndFutureTests.swift
@@ -31,57 +31,6 @@ class PromiseAndFutureTestsTests: XCTestCase {
     // MARK: -
 
     // MARK: completing
-    
-    func testX() {
-        var promise: Promise<Int> = .init()
-        
-        var e = promise.future.expectValue(timeout: 1)
-        promise.future.expectValue({ $0 == 1 }, timeout: 1)
-        promise.future.expectError(timeout: 1)
-        promise.future.expectError({ $0 is String }, timeout: 1)
-        promise.future.expectCancelled(timeout: 1)
-        e.waitForExpectation()
-        
-        promise = .init()
-        promise.fulfill(with: 0)
-        
-        e = promise.future.expectValue(timeout: 1)
-        promise.future.expectValue({ $0 == 1 }, timeout: 1)
-        promise.future.expectError(timeout: 1)
-        promise.future.expectError({ $0 is String }, timeout: 1)
-        promise.future.expectCancelled(timeout: 1)
-        e.waitForExpectation()
-        
-        promise = .init()
-        promise.fulfill(with: 1)
-        
-        e = promise.future.expectValue(timeout: 1)
-        promise.future.expectValue({ $0 == 1 }, timeout: 1)
-        promise.future.expectError(timeout: 1)
-        promise.future.expectError({ $0 is String }, timeout: 1)
-        promise.future.expectCancelled(timeout: 1)
-        e.waitForExpectation()
-        
-        promise = .init()
-        promise.break(with: testError)
-        
-        e = promise.future.expectValue(timeout: 1)
-        promise.future.expectValue({ $0 == 1 }, timeout: 1)
-        promise.future.expectError(timeout: 1)
-        promise.future.expectError({ $0 is String }, timeout: 1)
-        promise.future.expectCancelled(timeout: 1)
-        e.waitForExpectation()
-        
-        promise = .init()
-        promise.cancel()
-        
-        e = promise.future.expectValue(timeout: 1)
-        promise.future.expectValue({ $0 == 1 }, timeout: 1)
-        promise.future.expectError(timeout: 1)
-        promise.future.expectError({ $0 is String }, timeout: 1)
-        promise.future.expectCancelled(timeout: 1)
-        e.waitForExpectation()
-    }
 
     func testShouldHandleValue_WhenCompletingWithValue() {
         promise.future
